Unconscious Timer increased
Before it would kick them from game roughly 2 minutes to log back in,
now its 3 minutes in game unconscious so makes at least 1 friend to keep
watch for zombies players then a ton of people just trying to jailbreak
a door. So more risk and still able to do it.